+++ components/login/passwordReset/test/index.test.js
@@ -6,40 +6,5 @@
 /* eslint no-undef:0 */
 
 import '../src/domain/test/index.js'
-
-import ReactDOM from 'react-dom'
-
-import chai, {expect} from 'chai'
-import chaiDOM from 'chai-dom'
-
-import Component from '../src/index.js'
-chai.use(chaiDOM)
-
-describe('LoginPasswordReset', () => {
-  const setup = setupEnvironment(Component)
-
-  it('should render without crashing', () => {
-    // Given
-    const props = {}
-
-    // When
-    const component = <Component {...props} />
-
-    // Then
-    const div = document.createElement('div')
-    ReactDOM.render(component, div)
-    ReactDOM.unmountComponentAtNode(div)
-  })
-
-  it('should NOT render null', () => {
-    // Given
-    const props = {}
-
-    // When
-    const {container} = setup(props)
-
-    // Then
-    expect(container.innerHTML).to.be.a('string')
-    expect(container.innerHTML).to.not.have.lengthOf(0)
-  })
-})
+import './resetPasswordForm.test.js'
+import './changePasswordForm.test.js'