+++ components/login/passwordReset/test/utils.js
@@ -1,0 +1,6 @@
+export const setTokenAndExpiration = (token, exp, global) => {
+  const url = new URL(global.window.location)
+  url.searchParams.set('token', token)
+  url.searchParams.set('exp', exp)
+  global.window.history.pushState(null, '', url.toString())
+}